requirements.txt
# Data and Analysis
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
ta==0.10.2
yfinance==0.2.24

# Crypto APIs
ccxt==4.1.74
python-binance==1.0.17
requests==2.31.0

# Database
sqlite3

# Backtesting
zipline-reloaded==3.0.3
pyfolio==0.9.2

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6

# Async
asyncio
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Plotting and Visualization
matplotlib==3.8.2
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
schedule==1.2.0
python-dateutil==2.8.2